Use integer for slice indexes

+++ src/main/java/be/cytomine/client/models/AbstractSlice.java
@@ -4,11 +4,11 @@ public class AbstractSlice extends Model<AbstractSlice> {
 
     public AbstractSlice() {}
 
-    public AbstractSlice(AbstractImage image, UploadedFile uploadedFile, String mime, Double channel, Double zStack, Double time) {
+    public AbstractSlice(AbstractImage image, UploadedFile uploadedFile, String mime, Integer channel, Integer zStack, Integer time) {
         this(image.getId(), uploadedFile.getId(), mime, channel, zStack, time);
     }
 
-    public AbstractSlice(Long imageId, Long uploadedFileId, String mime, Double channel, Double zStack, Double time) {
+    public AbstractSlice(Long imageId, Long uploadedFileId, String mime, Integer channel, Integer zStack, Integer time) {
         this.set("image", imageId);
         this.set("uploadedFile", uploadedFileId);
         this.set("mime", mime);
